+++ src/onegov.core/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds the ability to version static files to cache them forever.
+  [href]
+
 0.70.5 (2018-06-04)
 ~~~~~~~~~~~~~~~~~~~